% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_books.R
\name{combine_books}
\alias{combine_books}
\title{combine_books}
\usage{
combine_books(books, fmt = NULL)
}
\arguments{
\item{books}{a simple table}

\item{fmt}{a format, defaults to NULL}
}
\value{
dataframe
}
\description{
combine_books
}
