% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_sv.R
\name{fbmet_sv}
\alias{fbmet_sv}
\title{fbmet_sv}
\usage{
fbmet_sv(input, output, session, values)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{values}{shiny values}
}
\description{
fbmet_sv
}
