% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_biplot.R
\name{gg_biplot}
\alias{gg_biplot}
\title{gg_biplot}
\usage{
gg_biplot(model, add_labels = TRUE, distance = 0.5, ranges = NULL)
}
\arguments{
\item{model}{a model as from the agricolae function AMMI}

\item{add_labels}{whether the biplot should show labels (more time consuming)}

\item{distance}{the relative distance at which to draw the circle, default: 0.5}

\item{ranges}{give xlim and ylim ranges in plot units; default is NULL. Allows re-
sizing.}
}
\description{
Creates a modified biplot. The modification is re-sizing dots by trait value.
}
\examples{
# Get sample data
library(agricolae)
data(plrv)

library(fbmet)

model<- with(plrv, AMMI(Locality, Genotype, Rep, Yield, console=FALSE))

if(interactive()){
  gg_biplot(model)
}

}

