% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_env_plot.R
\name{gg_env_plot}
\alias{gg_env_plot}
\title{gg_env_plot}
\usage{
gg_env_plot(atable, genotypes = unique(as.character(atable$Genotype))[1:10],
  trait = names(atable)[ncol(atable)])
}
\arguments{
\item{atable}{a dataframe containing at least three columns: Genotype, Locality, [Trait]
in this order.}

\item{genotypes}{a subset of up to 10 genotypes that will be highlighted through lines.}

\item{trait}{the trait (or index) to be shown}
}
\description{
Shows performance of a trait or index for all varties separately in each environment.
Data for each environment are displayed as violin plots.
}
\details{
One or more (up to 10) genotypes can be highlighted, that is they will be displayed in
differntly colored dots and are connected through lines which are also correspondingly
colored.
}
\examples{
# Get sample data
library(agricolae)
data(plrv)

library(fbmet)

# enable easy summaries
#library(magrittr)
library(dplyr)

atable <- plrv \%>\% group_by(Genotype, Locality) \%>\% summarise(Yield = mean(Yield))

if(interactive()){
  gg_env_plot(atable)
}

}

