% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_hist.R
\name{gg_hist}
\alias{gg_hist}
\title{gg_hist}
\usage{
gg_hist(model, filtered = unique(row.names(model$biplot)))
}
\arguments{
\item{model}{as from AMMI (agricolae library)}

\item{filtered}{a vector of filtered genotypes.}
}
\description{
Shows histogram of averages.
}

