% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedBiplotUI.R
\name{linkedBiplotUI}
\alias{linkedBiplotUI}
\title{linkedBiplotUI}
\usage{
linkedBiplotUI(id)
}
\arguments{
\item{id}{shiny}
}
\value{
fluidRow (shiny fragment)
}
\description{
linkedBiplotUI
}
\author{
Reinhard Simon
}

