% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_selected.R
\name{met_selected}
\alias{met_selected}
\title{met_selected}
\usage{
met_selected(input, output, session, model, ndat)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{model}{AMMI model}

\item{ndat}{average data by genotype and environment}
}
\description{
Select interactively genotypes from a multi-environmental dataset.
}

