% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRData.R
\name{loadRData}
\alias{loadRData}
\title{Title}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{path to RData fiel}
}
\value{
R object
}
\description{
Title
}
\author{
ricardo
}
\references{
\url{http://stackoverflow.com/questions/5577221/how-can-i-load-an-object-into-a-variable-name-that-i-specify-from-an-r-data-file}
}

